requirements.txt
pandas
numpy
scikit-learn
shap
sqlalchemy
mlflow
loguru
python-dotenv
flask
reportlab
flask-mail